changes to folder structure:
made changes to the folder structure, created the api folder which contains views, urls and serializers.

changes to models:

made changes to the relations from many to one to many to many.

changes to the views:

made changes to the views to show multiple countries,authors and publishers in the book-list